+++ src/store/assetInfoStore.ts
@@ -11,7 +11,12 @@ import {
 } from "@gardenfi/orderbook";
 import { API } from "../constants/api";
 import axios from "axios";
-import { Quote, Strategies } from "@gardenfi/core";
+import {
+  BitcoinNetwork,
+  BitcoinProvider,
+  Quote,
+  Strategies,
+} from "@gardenfi/core";
 import { generateTokenKey } from "../utils/generateTokenKey";
 import { PublicKey } from "@solana/web3.js";
 
@@ -27,6 +32,7 @@ import {
   getStarknetTokenBalance,
 } from "../utils/getTokenBalance";
 import { Hex } from "viem";
+import { getSpendableBalance } from "../utils/getmaxBtc";
 
 export type Networks = {
   [chain in Chain]: ChainData & { assetConfig: Omit<AssetConfig, "chain">[] };
@@ -84,7 +90,8 @@ type AssetInfoState = {
     fetchOnlyAsset?: Asset
   ) => Promise<void>;
   fetchAndSetBitcoinBalance: (
-    provider: IInjectedBitcoinProvider
+    provider: IInjectedBitcoinProvider,
+    address: string
   ) => Promise<void>;
   fetchAndSetStarknetBalance: (address: string) => Promise<void>;
   fetchAndSetSolanaBalance: (address: PublicKey) => Promise<void>;
@@ -275,7 +282,11 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
     }
   },
 
-  fetchAndSetBitcoinBalance: async (provider: IInjectedBitcoinProvider) => {
+  fetchAndSetBitcoinBalance: async (
+    provider: IInjectedBitcoinProvider,
+
+    address: string
+  ) => {
     const { assets } = get();
     if (!assets || !provider) return;
 
@@ -284,12 +295,28 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
       if (!balance?.val?.total) return;
 
       const formattedBalance = new BigNumber(balance.val.confirmed);
+
+      const _provider = new BitcoinProvider(
+        network === "mainnet" ? BitcoinNetwork.Mainnet : BitcoinNetwork.Testnet
+      );
+
+      const feeRate = await _provider.getFeeRates();
+      const utxos = await _provider.getUTXOs(address, Number(formattedBalance));
+      const spendable = await getSpendableBalance(
+        address,
+        Number(formattedBalance),
+        utxos.length,
+        feeRate.fastestFee
+      );
+      const maxSpendableBalance = spendable.ok ? spendable.val : 0;
+
       const btcBalance = Object.values(assets)
         .filter((asset) => isBitcoin(asset.chain))
         .reduce(
           (acc, asset) => {
-            acc[getOrderPair(asset.chain, asset.tokenAddress)] =
-              formattedBalance;
+            acc[getOrderPair(asset.chain, asset.tokenAddress)] = new BigNumber(
+              maxSpendableBalance
+            );
             return acc;
           },
           {} as Record<string, BigNumber | undefined>