+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
-## [1.0.8]
+## [1.0.10]
 
 ### Added
 
-- Added support for solana spl tokens in staging
-- Added support for solana spl tokens in testnet
+- Added support for solana spl tokens in testnet.
+
+## [1.0.9]
+
+### Fixed
+
+- Fetch and set max spendable balance for Bitcoin and Solana at the time of fetching balances.
+
+## [1.0.8]
+
+### Fixed
+
+- Fixed posting refundSacp for an order regardless any state
 
 ## [1.0.7]